+++ src/Models/Account.ts
@@ -63,7 +63,7 @@ class Account {
             }
         });
 
-        checkStatus(response);
+        await checkStatus(response);
 
         const respData = await response.json();
 
@@ -128,7 +128,7 @@ class Account {
             }
         });
 
-        checkStatus(response);
+        await checkStatus(response);
 
         const respData = await response.json();
 
@@ -157,15 +157,14 @@ class Account {
                 message: 'User is not authenticated'
             });
         }
-        const t = this.getToken();
 
         const response = await fetch(this.config.API.signOut, {
             headers: {
-                Authorization: 'Bearer ' + t.token
+                Authorization: `Bearer ${this.getToken().token}`
             }
         });
 
-        checkStatus(response);
+        await checkStatus(response);
 
         this.setToken('', '', 0);
 
@@ -180,9 +179,7 @@ class Account {
  * @return {Boolean}
  */
     isAuthenticated() {
-        const t = this.getToken();
-
-        return !t.isExpired() && t.token !== '';
+        return !this.getToken().isExpired() && this.getToken().token !== '';
     }
 
     /** Retruns the OAuth token
@@ -255,7 +252,8 @@ class Account {
             }
         });
 
-        checkStatus(response);
+        await checkStatus(response);
+
         const responseData = await response.json();
 
         this.setToken(
@@ -285,7 +283,7 @@ class Account {
             credentials: 'include'
         });
 
-        checkStatus(response);
+        await checkStatus(response);
 
         return await response.json();
     }
@@ -324,7 +322,7 @@ class Account {
             }
         });
 
-        checkStatus(response);
+        await checkStatus(response);
 
         return await response.json();
     }
@@ -362,7 +360,7 @@ class Account {
             }
         });
 
-        checkStatus(response);
+        await checkStatus(response);
 
     // Response is 204: No Content, nothing to return.
     }
@@ -385,15 +383,13 @@ class Account {
             });
         }
 
-        const t = this.getToken();
-
         const response = await fetch(this.config.API.getAccountInfo, {
             headers: {
-                Authorization: 'Bearer ' + t.token
+                Authorization: `Bearer ${this.getToken().token}`
             }
         });
 
-        checkStatus(response);
+        await checkStatus(response);
 
         return await response.json();
     }
@@ -414,7 +410,7 @@ class Account {
     async getSocialLoginUrls(state: any) {
         const response = await fetch(this.config.API.getSocialLoginUrls(state));
 
-        checkStatus(response);
+        await checkStatus(response);
 
         return await response.json();
     }
@@ -437,7 +433,6 @@ class Account {
                 message: 'User is not authenticated'
             });
         }
-        const t = this.getToken();
 
         let body: any = {
             full_name: data.fullName
@@ -451,12 +446,12 @@ class Account {
             method: 'PUT',
             body: params(body),
             headers: {
-                Authorization: 'Bearer ' + t.token,
+                Authorization: `Bearer ${this.getToken().token}`,
                 'Content-Type': 'application/x-www-form-urlencoded'
             }
         });
 
-        checkStatus(response);
+        await checkStatus(response);
 
         return await response.json();
     }
@@ -490,7 +485,6 @@ class Account {
                 message: 'User is not authenticated'
             });
         }
-        const t = this.getToken();
 
         let body = {
             old_password: data.oldPassword,
@@ -503,12 +497,12 @@ class Account {
             method: 'POST',
             body: params(body),
             headers: {
-                Authorization: 'Bearer ' + t.token,
+                Authorization: `Bearer ${this.getToken().token}`,
                 'Content-Type': 'application/x-www-form-urlencoded'
             }
         });
 
-        checkStatus(response);
+        await checkStatus(response);
 
         return await response.json();
     }
@@ -529,7 +523,7 @@ class Account {
             this.config.API.getRegisterFields(merchantUuid)
         );
 
-        checkStatus(response);
+        await checkStatus(response);
 
         return await response.json();
     }
@@ -558,7 +552,6 @@ class Account {
                 message: 'User is not authenticated'
             });
         }
-        const t = this.getToken();
 
         let body = {
             password: data.password,
@@ -568,13 +561,13 @@ class Account {
         const response = await fetch(this.config.API.deleteAccount, {
             method: 'DELETE',
             headers: {
-                Authorization: 'Bearer ' + t.token,
+                Authorization: `Bearer ${this.getToken().token}`,
                 'Content-Type': 'application/x-www-form-urlencoded'
             },
             body: params(body)
         });
 
-        checkStatus(response);
+        await checkStatus(response);
 
         localStorage.removeItem(this.config.INPLAYER_TOKEN_NAME);
         localStorage.removeItem(this.config.INPLAYER_IOT_NAME);
@@ -610,8 +603,6 @@ class Account {
             });
         }
 
-        const t = this.getToken();
-
         let body = {
             password: data.password,
             branding_id: data.brandingId
@@ -620,13 +611,13 @@ class Account {
         const response = await fetch(this.config.API.exportData, {
             method: 'POST',
             headers: {
-                Authorization: 'Bearer ' + t.token,
+                Authorization: `Bearer ${this.getToken().token}`,
                 'Content-Type': 'application/x-www-form-urlencoded'
             },
             body: params(body)
         });
 
-        checkStatus(response);
+        await checkStatus(response);
 
         if (!response.ok) {
             return await response.json();